refactor generate test dataset w/o env into common
