+++ tests/utils/test_dataset_combine.py
@@ -1,13 +1,16 @@
 from typing import Optional
 
 import gymnasium as gym
+import numpy as np
 import pytest
+from gymnasium import spaces
 from gymnasium.utils.env_checker import data_equivalence
 from packaging.specifiers import SpecifierSet
 
 import minari
 from minari import DataCollectorV0, MinariDataset
 from minari.utils import combine_datasets, combine_minari_version_specifiers
+from tests.common import get_sample_buffer_for_dataset_from_env
 
 
 def _check_env_recovery(gymnasium_environment: gym.Env, dataset: MinariDataset):
@@ -98,6 +101,52 @@ def _generate_dataset_with_collector_env(
         code_permalink="https://github.com/Farama-Foundation/Minari/blob/f095bfe07f8dc6642082599e07779ec1dd9b2667/tutorials/LocalStorage/local_storage.py",
         author="WillDudley",
         author_email="wdudley@farama.org",
+    )
+    assert isinstance(dataset, MinariDataset)
+    env.close()
+
+
+def _generate_dataset_without_env(dataset_id: str, num_episodes: int = 10):
+    """Helper function to create tmp dataset without an env to use for testing combining.
+
+    Args:
+        dataset_id (str): name of the generated Minari dataset
+        num_episodes (int): number of episodes in the generated dataset
+    """
+    buffer = []
+    action_space_subset = spaces.Dict(
+        {
+            "component_2": spaces.Dict(
+                {
+                    "subcomponent_2": spaces.Box(low=4, high=5, dtype=np.float32),
+                }
+            ),
+        }
+    )
+    observation_space_subset = spaces.Dict(
+        {
+            "component_2": spaces.Dict(
+                {
+                    "subcomponent_2": spaces.Box(low=4, high=5, dtype=np.float32),
+                }
+            ),
+        }
+    )
+
+    env = gym.make("DummyDictEnv-v0")
+    buffer = get_sample_buffer_for_dataset_from_env(env, num_episodes)
+
+    # Create Minari dataset and store locally
+    dataset = minari.create_dataset_from_buffers(
+        dataset_id=dataset_id,
+        buffer=buffer,
+        env=None,
+        algorithm_name="random_policy",
+        code_permalink="https://github.com/Farama-Foundation/Minari/blob/f095bfe07f8dc6642082599e07779ec1dd9b2667/tutorials/LocalStorage/local_storage.py",
+        author="WillDudley",
+        author_email="wdudley@farama.org",
+        action_space=action_space_subset,
+        observation_space=observation_space_subset,
     )
     assert isinstance(dataset, MinariDataset)
     env.close()
